Circumvent bug in Cipher.init() in openJDK-8
When Cipher.init() is called with an invalid key (causing an
InvalidKeyException), the iterator of SPIs is not reset. A subsequent call
to Cipher.init() on the same object with a valid key then still throws an
InvalidKeyException stating "No installed provider supports this key",
although this is not correct. Therefore, a new Cipher object is required
to find a supporting provider.